[FE] refactor: 생년월일 입력 폼 관련 상태 및 Ref 훅으로 이동 (#29)
