+++ src/frontend/src/components/common/AuthDateInput/useDateInput.ts
@@ -1,3 +1,5 @@
+import { useRef, useState } from 'react';
+
 const useDateInput = () => {
   const MAX_AGE = 150;
   const MAX_MONTH = 12;
@@ -8,7 +10,40 @@ const useDateInput = () => {
   const months = Array.from({ length: MAX_MONTH }, (_, index) => String(index + 1) + '월');
   const days = Array.from({ length: MAX_DAY }, (_, index) => String(index + 1));
 
-  return { years, months, days };
+  const [selectedYear, setSelectedYear] = useState('');
+  const [selectedMonth, setSelectedMonth] = useState('');
+  const [selectedDay, setSelectedDay] = useState('');
+
+  const monthRef = useRef<HTMLInputElement>(null);
+  const dayRef = useRef<HTMLInputElement>(null);
+
+  const handleYearSelect = (year: string) => {
+    setSelectedYear(year);
+    monthRef.current?.focus();
+  };
+
+  const handleMonthSelect = (month: string) => {
+    setSelectedMonth(month);
+    dayRef.current?.focus();
+  };
+
+  const handleDaySelect = (day: string) => {
+    setSelectedDay(day);
+  };
+
+  return {
+    years,
+    months,
+    days,
+    selectedYear,
+    selectedMonth,
+    selectedDay,
+    monthRef,
+    dayRef,
+    handleYearSelect,
+    handleMonthSelect,
+    handleDaySelect,
+  };
 };
 
 export default useDateInput;
